Consolidate DrugMapper and OrderMapper versions
