--- api/src/main/java/org/openmrs/module/emrapi/encounter/mapper/DrugMapper1_12.java
@@ -1,37 +1,0 @@
-/**
- * The contents of this file are subject to the OpenMRS Public License
- * Version 1.0 (the "License"); you may not use this file except in
- * compliance with the License. You may obtain a copy of the License at
- * http://license.openmrs.org
- *
- * Software distributed under the License is distributed on an "AS IS"
- * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
- * License for the specific language governing rights and limitations
- * under the License.
- *
- * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
- */
-package org.openmrs.module.emrapi.encounter.mapper;
-
-
-import org.openmrs.Drug;
-import org.openmrs.annotation.OpenmrsProfile;
-import org.openmrs.module.emrapi.encounter.DrugMapper;
-import org.openmrs.module.emrapi.encounter.domain.EncounterTransaction;
-import org.springframework.stereotype.Component;
-
-@Component(value = "drugMapper")
-public class DrugMapper1_12 implements DrugMapper {
-
-    public EncounterTransaction.Drug map(Drug drug){
-        EncounterTransaction.Drug encounterTransactionDrug = new EncounterTransaction.Drug();
-        encounterTransactionDrug.setName(drug.getDisplayName());
-        if (drug.getDosageForm() != null) {
-            encounterTransactionDrug.setForm(drug.getDosageForm().getName().getName());
-        }
-        encounterTransactionDrug.setStrength(drug.getStrength());
-        encounterTransactionDrug.setUuid(drug.getUuid());
-        return encounterTransactionDrug;
-    }
-
-}
